// src/enums/enums.ts
export enum Role {
  admin = 'ADMIN',
  proDirector = 'PRO_DIRECTOR',
  applicant = 'APPLICANT',
}

export enum Talent {
  ACTOR = 'ACTOR',
  PRODUCER = 'PRODUCER',
  DIRECTOR = 'DIRECTOR',
  SCREEN_WRITER = 'SCREEN WRITER',
  EXECUTIVE_PRODUCER = 'EXECUTIVE PRODUCER',
  PRODUCTION_MANAGER = 'PRODUCTION MANAGER',
  SCRIPT_SUPERVISOR = 'SCRIPT SUPERVISOR',
  LOCATION_MANAGER = 'LOCATION MANAGER',
  LOCATION_SCOUT = 'LOCATION SCOUT',
  CINEMATOGRAPHER = 'CINEMATOGRAPHER',
  CAMERA_OPERATOR = 'CAMERA OPERATOR',
  BOOM_OPERATOR = 'BOOM OPERATOR',
  GRIP = 'GRIP',
  GAFFER = 'GAFFER',
  BEST_BOY = 'BEST BOY',
  PRODUCTION_DESIGNER = 'PRODUCTION DESIGNER',
  ART_DIRECTOR = 'ART DIRECTOR',
  MAKE_UP_ARTIST = 'MAKE-UP ARTIST',
  COSTUME = 'COSTUME',
  EDITOR = 'EDITOR',
  COMPOSER = 'COMPOSER',
}

export enum Verification {
  FAMOUS = 'FAMOUS',
  ENDORSED = 'ENDORSED',
}

export enum Gender {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
}

export enum Status {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
}

export enum Language {
  AMHARIC = 'Amharic',
  AFAAN_OROMO = 'Afaan Oromo',
  TIGRIGNA = 'Tigrigna',
  SIDAMIGNA = 'Sidamigna',
  WOLAYTIGNA = 'Wolaytigna',
  ENGLISH = 'English',
  MANDARIN = 'Mandrin',
  SPANISH = 'Spanish',
  DEUTCH = 'Deutch',
}